moved useStates and relevant handler functions to Event and NewEventForm components
