+++ src/App.jsx
@@ -1,96 +1,86 @@
 import { useState } from "react";
 import eventsData from "./data";
-import { v1 as generateUniqueID } from "uuid";
-// import Attendees from "./Attendees";
+
+//import Attendee from "./Components/Attendee";
 import Event from "./Components/Event";
-// import Footer from "./Components/Footer";
-// import Header from "./Components/Header";
+import Footer from "./Components/Footer";
+import Header from "./Components/Header";
 import NewEventForm from "./Components/NewEventForm";
 
 function App() {
   
   const [events, setEvents] = useState(eventsData); // const [events, setEvents] = useState(eventsData);: Initializes events state with eventsData. This state holds the array of event objects.
 
-  const [showAttendees, setShowAttendees] = useState(false); // This controls whether the attendees of an event are displayed.
+  // sent to Events.jsx
+  // const [showAttendees, setShowAttendees] = useState(false); // This controls whether the attendees of an event are displayed.
 
-  const [selectOption, setSelectOption] = useState(""); // const [selectOption, setSelectOption] = useState("");: Initializes selectOption state to an empty string. It holds the selected option for the event type.
+  // NewEventForm.jsx
+  // const [selectOption, setSelectOption] = useState(""); // const [selectOption, setSelectOption] = useState("");: Initializes selectOption state to an empty string. It holds the selected option for the event type.
 
-  const [newEvent, setNewEvent] = useState({
-    id: "",
-    eventType: "",
-    name: "",
-    organizer: "",
-    eventImage: "",
-    date: "",
-    people: [],
-  });
+  // const [newEvent, setNewEvent] = useState({
+  //   id: "",
+  //   eventType: "",
+  //   name: "",
+  //   organizer: "",
+  //   eventImage: "",
+  //   date: "",
+  //   people: [],
+  // });
 
 
+// moved to NewEventForm.jsx
+  // function handleSelectChange(e) {
+  //   setSelectOption(e.target.value);
+  // }
+  // function handleSubmit(e) {
+  //   e.preventDefault();
+  //   addEvent();
+  //   resetEventForm();
+  // }
 
-  function handleSelectChange(e) {
-    setSelectOption(e.target.value);
-  }
-  function handleSubmit(e) {
-    e.preventDefault();
-    addEvent();
-    resetEventForm();
-  }
+  // sent to the NewEventForm.jsx
+  // function handleTextChange(e) {
+  //   setNewEvent({
+  //     ...newEvent,
+  //     [e.target.id]: e.target.value,
+  //   });
+  // }
 
-  function handleTextChange(e) {
-    setNewEvent({
-      ...newEvent,
-      [e.target.id]: e.target.value,
-    });
-  }
 
-  function resetEventForm() {
-    setNewEvent({
-      id: "",
-      eventType: "",
-      name: "",
-      organizer: "",
-      eventImage: "",
-      date: "",
-    });
-    setSelectOption("");
-  }
-  function resetEventForm() {
-    setNewEvent({
-      id: "",
-      eventType: "",
-      name: "",
-      organizer: "",
-      eventImage: "",
-      date: "",
-    });
-    setSelectOption("");
-  }
+  // function resetEventForm() {
+  //   setNewEvent({
+  //     id: "",
+  //     eventType: "",
+  //     name: "",
+  //     organizer: "",
+  //     eventImage: "",
+  //     date: "",
+  //   });
+  //   setSelectOption("");
+  // }
 
-  function handleSubmit(e) {
-    e.preventDefault();
-    addEvent();
-    resetEventForm();
-  }
-  function addEvent() {
-    const createEvent = {
-      id: generateUniqueID(),
-      eventType: selectOption,
-      name: newEvent.name,
-      organizer: newEvent.organizer,
-      eventImage: newEvent.eventImage || null,
-      date: newEvent.date,
-      people: [],
-    };
-    handleAddEvent(createEvent);
-  }
+
+  // sent to NewEventForm.jsx
+  // function addEvent() {
+  //   const createEvent = {
+  //     id: generateUniqueID(),
+  //     eventType: selectOption,
+  //     name: newEvent.name,
+  //     organizer: newEvent.organizer,
+  //     eventImage: newEvent.eventImage || null,
+  //     date: newEvent.date,
+  //     people: [],
+  //   };
+  //   handleAddEvent(createEvent);
+  // }
 
   function handleAddEvent(event) {
     setEvents([event, ...events]);
   }
-
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
-  }
+// sent to Events.jsx
+  // function toggleEventAttendees() {
+  //   setShowAttendees(!showAttendees);
+  // }
 
   function updateEventAttendance(eventId, attendeeId) {
     const eventArray = [...events];
@@ -109,28 +99,21 @@ function App() {
   return (
     <div className="App">
       <>
-        <header>
-          <h1 className="color-change-5x">RSVP App</h1>
-        </header>
+        <Header />
       </>
       <main>
         <NewEventForm
-        handleSubmit={handleSubmit} handleTextChange= {handleTextChange} newEvent = {newEvent} handleSelectChange={handleSelectChange} />
+        addEvent={addEvent} handleSubmit={handleSubmit} handleTextChange= {handleTextChange} newEvent = {newEvent} handleSelectChange={handleSelectChange} />
           <div className="events">
-          {/* <ul> */}
-          <Event events={events} updateEventAttendance={updateEventAttendance} showAttendees={showAttendees} toggleEventAttendees={toggleEventAttendees}/>
-          {/* </ul> */}
+          <ul>
+          {events.map((event) => {
+              const { people: attendees } = event;
+              return <Event key={event.id} events={events} attendees={attendees} updateEventAttendance={updateEventAttendance}/>
+            })}
+          </ul>
           </div>
       </main>
-      <>
-        <footer>
-          <ul>
-            <li>Contact</li>
-            <li>About</li>
-            <li>Legal</li>
-          </ul>
-        </footer>
-      </>
+        <Footer />
     </div>
   );
 }
